% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_disassociate_configuration_items_from_application}
\alias{applicationdiscoveryservice_disassociate_configuration_items_from_application}
\title{Disassociates one or more configuration items from an application}
\usage{
applicationdiscoveryservice_disassociate_configuration_items_from_application(
  applicationConfigurationId, configurationIds)
}
\arguments{
\item{applicationConfigurationId}{[required] Configuration ID of an application from which each item is
disassociated.}

\item{configurationIds}{[required] Configuration ID of each item to be disassociated from an application.}
}
\description{
Disassociates one or more configuration items from an application.
}
\section{Request syntax}{
\preformatted{applicationdiscoveryservice$disassociate_configuration_items_from_application(
  applicationConfigurationId = "string",
  configurationIds = list(
    "string"
  )
)
}
}

\keyword{internal}
