% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_start_continuous_export}
\alias{applicationdiscoveryservice_start_continuous_export}
\title{Start the continuous flow of agent's discovered data into Amazon Athena}
\usage{
applicationdiscoveryservice_start_continuous_export()
}
\description{
Start the continuous flow of agent's discovered data into Amazon
Athena.
}
\section{Request syntax}{
\preformatted{applicationdiscoveryservice$start_continuous_export()
}
}

\keyword{internal}
